extract onRenderValueLabels into another file
